import the rest of the work, seems to be running solid.
reduced things down to the xrandr approach (using a command)
and just one Screen entry for each pair of video outputs.
they are tied together with a Virtual setting for the width
and the xrandr command shuffles the DVI outputs around appropriately.

until the xrandr command, the screens just represent a clone of each other.